Don't download the core when a local build is present
There's no particularly pleasant way to set REALM_CORE_VERSION when
building from Xcode, so try to automatically detect that the copy of the
core was built locally and leave it alone if so.

+++ build.sh
@@ -178,6 +178,10 @@ case "$COMMAND" in
             exit 0
         fi
         if ! [ -L core ]; then
+            if [ -d core -a -d ../tightdb ]; then
+                echo "Local build of core appears to be present."
+                exit 0
+            fi
             echo "core is not a symlink. Deleting..."
             rm -rf core
             download_core
